pyproject.toml
[project]
name = "clip-similarity-matrix"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "torch>=1.9.0",
    "transformers>=4.21.0",
    "Pillow>=8.3.0",
    "matplotlib>=3.5.0",
    "numpy>=1.21.0",
    "seaborn>=0.11.0",
    "ruff",
    "pydantic",
    "scikit-learn>=1.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff.lint]
ignore = ["RUF012", "RUF002", "T201", "SIM105", "RET504"]
select = [
    "A",
    "B",
    "C4",
    "DTZ",
    "E",
    "F",
    "FAST",
    "G",
    "I",
    "ISC",
    "LOG",
    "N",
    "NPY",
    "PD",
    "RET",
    "RUF",
    "SIM",
    "T20",
    "TID",
    "UP",
    "W",
]

[tool.pyright]
pythonVersion = "3.12"

[tool.hatch.build.targets.wheel]
packages = ["utils", "scripts"]
